LayoutTests/platform/glib/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/reload-rejection-order-unload-unserializablestate-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_throws_dom: function "() => { throw committedReason; }" threw undefined with type "undefined", not an object


FAIL reload() with an unserializable state inside onunload throws "DataCloneError", not "InvalidStateError" assert_equals: expected 1 but got 0
